+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 All notable changes to this project will be documented in this file.
+### [0.9.20] - 2018-07-23
+[fix] zero value filtering
+[feature] message about start and end filtering
 ### [0.9.19] - 2018-07-20
 [fix] handling errors by MQTT
 ### [0.9.18] - 2018-07-17